requirements.txt
Telethon
pyenchant